+++ line.go
@@ -67,12 +67,12 @@ func (l Line[T]) IsZero() bool {
 	return l.Start.IsZero() && l.End.IsZero()
 }
 
-// Int converts the line to a [int] line.
+// Int converts the line to a Line[int].
 func (l Line[T]) Int() Line[int] {
 	return Line[int]{l.Start.Int(), l.End.Int()}
 }
 
-// Float converts the line to a [float64] line.
+// Float converts the line to a Line[float64].
 func (l Line[T]) Float() Line[float64] {
 	return Line[float64]{l.Start.Float(), l.End.Float()}
 }