+++ polygon.go
@@ -84,12 +84,12 @@ func (p Polygon[T]) Empty() bool {
 	return len(p.Vertices) == 0
 }
 
-// Int converts the polygon to a [int] polygon.
+// Int converts the polygon to a Polygon[int].
 func (p Polygon[T]) Int() Polygon[int] {
 	return Polygon[int]{slices.Map(p.Vertices, Point[T].Int)}
 }
 
-// Float converts the polygon to a [float64] polygon.
+// Float converts the polygon to a Polygon[float64].
 func (p Polygon[T]) Float() Polygon[float64] {
 	return Polygon[float64]{slices.Map(p.Vertices, Point[T].Float)}
 }